% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{Get path to epinionRWeighting example}
\usage{
example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
epinionRWeighting comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
example()
example("sample_test.xlsx")
}
