% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.R
\name{epinion_weighting_tool}
\alias{epinion_weighting_tool}
\title{Weigh data}
\usage{
epinion_weighting_tool(
  data_input,
  weight_matrix_file,
  unique_id_var,
  fileEncoding = "UTF-8"
)
}
\arguments{
\item{data_input}{the name of the file which the data are to be read from.}

\item{weight_matrix_file}{weight matrix (.txt file).}

\item{unique_id_var}{variable used to identify each case as a unique in the data.}

\item{fileEncoding}{character string declares the encoding used on the weight_matrix_file.}
}
\value{
A data with weight variable included
}
\description{
This function used to weigh the data so that it could meet the distribution of population's demographics.
}
\examples{
data_weighted <- epinion_weighting_tool(data_input = example("sample_test.sav"),
weight_matrix_file = example("Weight matrix.txt"),
unique_id_var = "RespId")
}
